Refactor all select2 fields to use a common base class
Working on getting rid of a lot of code duplication by makins all of
these select2-powered fields inherit from a common base class.
